+++ app/Filament/Resources/AdminAuditLogResource/Pages/ListAdminAuditLogs.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\AdminAuditLogResource\Pages;
+
+use App\Filament\Resources\AdminAuditLogResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListAdminAuditLogs extends ListRecords
+{
+    protected static string $resource = AdminAuditLogResource::class;
+}